+++ src/config/questConfig.js
@@ -24,6 +24,7 @@ const { shopItems } = require("./shopConfig");
  *  checkRequirements: (trainer: WithId<Trainer>) => Promise<boolean>,
  *  rewards: Rewards,
  *  image?: string,
+ *  tip?: string,
  * }} TutorialStageData
  */
 
@@ -51,9 +52,9 @@ const newTutorialConfigRaw = {
     name: "Catching Pokemon",
     emoji: emojis.POKEBALL,
     description:
-      "The first thing to do is catch Pokemon! **To catch new Pokemon, use the `/gacha` command**. The gacha costs Pokeballs, with rare Pokeballs catching rarer Pokemon. I would recommend using 10 Pokeballs on the standard banner.",
+      "The first thing to do is catch Pokemon! **To catch new Pokemon, use the `/gacha` command**. \n## `/gacha`",
     requirementString: "Catch 1x Pokemon",
-    proceedString: "Use `/gacha` and catch a Pokemon!",
+    proceedString: "Use `/gacha` and catch a Pokemon! \n## `/gacha`",
     checkRequirements: async (trainer) =>
       await checkNumPokemon(trainer)
         .then((res) => (res.numPokemon ?? 0) > 0)
@@ -63,14 +64,15 @@ const newTutorialConfigRaw = {
     },
     image:
       "https://raw.githubusercontent.com/ewei068/pokestar/main/media/images/gacha.gif",
+    tip: "Rarer Pokeballs catch rarer Pokemon!",
   },
   dailyRewards: {
     name: "Daily Rewards",
     emoji: "📅",
     description:
-      "You can get more money and Pokeballs for free every day! **Use `/daily` to claim your daily rewards.**",
+      "You can get more money and Pokeballs for free every day! **Use `/daily` to claim your daily rewards.** \n## `/daily`",
     requirementString: "Claim Daily Rewards",
-    proceedString: "Use `/daily` to claim your daily rewards!",
+    proceedString: "Use `/daily` to claim your daily rewards! \n## `/daily`",
     checkRequirements: async (trainer) => trainer.claimedDaily,
     rewards: {
       backpack: {
@@ -85,10 +87,11 @@ const newTutorialConfigRaw = {
   backpack: {
     name: "Backpack Management",
     emoji: "🎒",
-    description: "You can view your items and money by using `/backpack`.",
+    description:
+      "You can view your items and money by using `/backpack`. \n## `/backpack`",
     requirementString: "Complete the previous stage",
     proceedString:
-      "Use `/backpack` to view your backpack, and complete the previous stage.",
+      "Use `/backpack` to view your backpack, and complete the previous stage. \n## `/backpack`",
     checkRequirements: async (trainer) =>
       trainer.tutorialData.completedTutorialStages.dailyRewards,
     rewards: {
@@ -101,9 +104,9 @@ const newTutorialConfigRaw = {
     name: "Catch More Pokemon",
     emoji: emojis.GREATBALL,
     description:
-      "Catch more Pokemon to build your team! **Use `/gacha` to catch 6 Pokemon**.",
+      "Catch more Pokemon to build your team! **Use `/gacha` to catch 6 Pokemon**.\n## `/gacha`",
     requirementString: "Catch 6x Pokemon",
-    proceedString: "Use `/gacha` to catch 6 Pokemon!",
+    proceedString: "Use `/gacha` to catch 6 Pokemon! \n## `/gacha`",
     checkRequirements: async (trainer) =>
       await checkNumPokemon(trainer)
         .then((res) => (res.numPokemon ?? 0) >= 6)
@@ -122,10 +125,10 @@ const newTutorialConfigRaw = {
     name: "Building Your Party",
     emoji: emojis.PIKACHU,
     description:
-      "Build your party by adding 6 Pokemon to your party! **Use `/party auto` to add your 6 strongest Pokemon to your party**. You can also use `/party manage` to make finer adjustments to your party.",
+      "Build your party by adding 6 Pokemon to your party! **Use `/party auto` to add your 6 strongest Pokemon to your party**. You can also use `/party manage` to make finer adjustments to your party. \n## `/party auto` `/party manage`",
     requirementString: "Add 6x Pokemon to Party",
     proceedString:
-      "Use `/party auto` or `/party manage` to add a Pokemon to your party!",
+      "Use `/party auto` or `/party manage` to add a Pokemon to your party! \n## `/party auto` `/party manage`",
     checkRequirements: async (trainer) => {
       const partyPokemon = trainer.party.pokemonIds.filter((id) => id);
       return partyPokemon.length >= 6;
@@ -144,7 +147,7 @@ const newTutorialConfigRaw = {
     name: "Learning to Battle: Taking Turns",
     emoji: "➡️",
     description:
-      "Before you battle, you must learn how they work! Battles in Pokestar are unique; **all 6 Pokemon fight at a time!**\n\n**Taking turns is based off combat readiness.** Pokemon with higher speed gain combat readiness faster. The current active Pokemon is highlighted in asterisks.\n\nYou may view the combat readiness of a team by clicking the **🔴 NPC** or **🔵 Player** tabs. There is also an indicator of which Pokemon moves next.",
+      "Before you battle, you must learn how they work! Battles in Pokestar are unique; **all 6 Pokemon fight at a time!**\n\n**Taking turns is based off combat readiness.** Pokemon with higher speed gain combat readiness faster. The current active Pokemon is highlighted in asterisks.",
     requirementString: "Complete the previous stage",
     proceedString:
       "Read the description to learn about battles, and complete the previous stage.",
@@ -155,6 +158,7 @@ const newTutorialConfigRaw = {
     },
     image:
       "https://raw.githubusercontent.com/ewei068/pokestar/main/media/images/tutorial/combat-readiness.png",
+    tip: "View the combat readiness of a team by clicking the 🔴 NPC or 🔵 Player tabs",
   },
   learnBattleMoves: {
     name: "Learning to Battle: Using Moves",
@@ -177,7 +181,7 @@ const newTutorialConfigRaw = {
     emoji: "🎯",
     description:
       "When using a move, you must select a target! **Click on the Pokemon you want to target from the dropdown menu.** Most moves may only target certain Pokemon, which is indicated in the **Target:** section of the move description." +
-      "\n\n**Some moves may affect an area of Pokemon.** When a target is selected, that area is indicated by a wider border. **When satisfied, click the ⚔️ Confirm button to use the move.** This can be disabled in your `/settings`." +
+      "\n\n**Some moves may affect an area of Pokemon.** When a target is selected, that area is indicated by a wider border. **When satisfied, click the ⚔️ Confirm button to use the move.**" +
       "\n\nFor more detailed battle mechanics, check out the [documentation on Github](https://github.com/ewei068/pokestar?tab=readme-ov-file#-battle-mechanics).",
     requirementString: "Complete the previous stage",
     proceedString:
@@ -189,14 +193,15 @@ const newTutorialConfigRaw = {
     },
     image:
       "https://raw.githubusercontent.com/ewei068/pokestar/main/media/images/tutorial/battle-target-select.gif",
+    tip: "You can disable target indicators in your /settings",
   },
   winNpcBattle: {
     name: "Battle an NPC!",
     emoji: "⚔️",
     description:
-      "Now that you know how to battle, **use `/pve` to battle an NPC!** I'd recommend starting with the <:bugcatcher:1117871382399815812> Bug Catcher on Very Easy difficulty.",
+      "Now that you know how to battle, **use `/pve` to battle an NPC!** I'd recommend starting with the <:bugcatcher:1117871382399815812> Bug Catcher on Very Easy difficulty. \n## `/pve`",
     requirementString: "Win any NPC Battle",
-    proceedString: "Use `/pve` to battle an NPC!",
+    proceedString: "Use `/pve` to battle an NPC! \n## `/pve`",
     checkRequirements: async (trainer) =>
       Object.keys(trainer.defeatedNPCs).length > 0,
     rewards: {
@@ -210,14 +215,15 @@ const newTutorialConfigRaw = {
     },
     image:
       "https://raw.githubusercontent.com/ewei068/pokestar/main/media/images/pve.gif",
+    tip: "Some NPCs give daily victory rewards!",
   },
   buyPokeballs: {
     name: "Buying Pokeballs",
     emoji: emojis.POKEBALL,
-    description: `One of the best ways to get more Pokeballs is to buy them from the Pokemart every day. Use \`/pokemart\` to buy ${emojis.POKEBALL} Pokeballs, or use \`/buy itemid: 0 quantity: 5\` to buy the maximum amount.`,
+    description: `One of the best ways to get more Pokeballs is to buy them from the Pokemart every day. Use \`/pokemart\` to buy ${emojis.POKEBALL} Pokeballs, or use \`/buy itemid: 0 quantity: 5\` to buy the maximum amount.\n## \`/pokemart\``,
     requirementString: `Buy 5x ${emojis.POKEBALL} Pokeballs`,
     proceedString:
-      "Use `/pokemart` or `/buy itemid: 0 quantity: 5` to buy 5 Pokeballs!",
+      "Use `/pokemart` or `/buy itemid: 0 quantity: 5` to buy 5 Pokeballs! \n## `/pokemart`",
     checkRequirements: async (trainer) =>
       trainer.purchasedShopItemsToday[shopItems.RANDOM_POKEBALL] >= 5,
     rewards: {
@@ -229,15 +235,16 @@ const newTutorialConfigRaw = {
     },
     image:
       "https://raw.githubusercontent.com/ewei068/pokestar/main/media/images/pokemart.gif",
+    tip: "You can buy 5 Pokeballs every day!",
   },
   listPokemon: {
     name: "Viewing Pokemon: List",
     emoji: "📜",
     description:
-      "You can view all your Pokemon using `/list`. **Use `/list` to view your Pokemon.**",
+      "You can view all your Pokemon using `/list`. **Use `/list` to view your Pokemon.** \n## `/list`",
     requirementString: "Complete the previous stage",
     proceedString:
-      "Use `/list` to view your Pokemon, and complete the previous stage.",
+      "Use `/list` to view your Pokemon, and complete the previous stage. \n## `/list`",
     checkRequirements: async (trainer) =>
       trainer.tutorialData.completedTutorialStages.winNpcBattle,
     rewards: {
@@ -245,15 +252,16 @@ const newTutorialConfigRaw = {
     },
     image:
       "https://raw.githubusercontent.com/ewei068/pokestar/main/media/images/tutorial/list.png",
+    tip: "You can search, sort, and release Pokemon from the /list!",
   },
   viewPokemon: {
     name: "Viewing Pokemon: Info",
     emoji: "🔍",
     description:
-      "You can view a specific Pokemon using `/info`, or clicking them from the `/list` dropdown menu. **Use `/info` to view a specific Pokemon.**\n\nFor **mobile users**, you can long press + copy to copy a Pokemon's ID from here.",
+      "You can view a specific Pokemon using `/info`, or clicking them from the `/list` dropdown menu. **Use `/info` to view a specific Pokemon.**\n## `/info <name>`",
     requirementString: "Complete the previous stage",
     proceedString:
-      "Use `/info` to view a specific Pokemon, and complete the previous stage.",
+      "Use `/info` to view a specific Pokemon, and complete the previous stage. \n## `/info <name>`",
     checkRequirements: async (trainer) =>
       trainer.tutorialData.completedTutorialStages.listPokemon,
     rewards: {
@@ -261,14 +269,16 @@ const newTutorialConfigRaw = {
     },
     image:
       "https://raw.githubusercontent.com/ewei068/pokestar/main/media/images/info.png",
+    tip: "For mobile users, you can long press + copy to copy a Pokemon's ID from here",
   },
   trainPokemon: {
     name: "Training Pokemon",
     emoji: "🏋️",
     description:
-      "One way to give your Pokemon more EXP is by training! Copy a Pokemon's ID, then **Use `/train` to train a Pokemon to level 12.**\n\nTip: On desktop, you can use the Up Arrow key to use your last command.",
+      "One way to give your Pokemon more EXP is by training! Copy a Pokemon's ID, then **Use `/train` to train a Pokemon to level 12.**\n## `/train <id>`",
     requirementString: "Have 1x Pokemon at level 12 or higher",
-    proceedString: "Use `/train` to train a Pokemon to level 12.",
+    proceedString:
+      "Use `/train` to train a Pokemon to level 12. \n## `/train <id>`",
     checkRequirements: async (trainer) => {
       const { data: pokemons } = await listPokemons(trainer, {
         pageSize: 1,
@@ -288,15 +298,16 @@ const newTutorialConfigRaw = {
     },
     image:
       "https://raw.githubusercontent.com/ewei068/pokestar/main/media/images/train.png",
+    tip: "On desktop, you can use the Up Arrow key to use your last command.",
   },
   trainerInfo: {
     name: "Viewing Trainer Info",
     emoji: "📊",
     description:
-      "You can view your trainer info using `/trainerinfo`. **Use `/trainerinfo` to view your trainer info.**",
+      "You can view your trainer info using `/trainerinfo`. **Use `/trainerinfo` to view your trainer info.** \n## `/trainerinfo`",
     requirementString: "Complete the previous stage",
     proceedString:
-      "Use `/trainerinfo` to view your trainer info, and complete the previous stage.",
+      "Use `/trainerinfo` to view your trainer info, and complete the previous stage. \n## `/trainerinfo`",
     checkRequirements: async (trainer) =>
       trainer.tutorialData.completedTutorialStages.trainPokemon,
     rewards: {
@@ -304,6 +315,7 @@ const newTutorialConfigRaw = {
     },
     image:
       "https://raw.githubusercontent.com/ewei068/pokestar/main/media/images/trainerinfo.png",
+    tip: "You can also view the trainer info of other trainers",
   },
   trainerLevel: {
     name: "Trainer Level",
@@ -326,9 +338,10 @@ const newTutorialConfigRaw = {
     name: "Level Rewards",
     emoji: "🎁",
     description:
-      "Every time you level up, you gain level rewards! **Use `/levelrewards` to claim your level rewards.**",
+      "Every time you level up, you gain level rewards! **Use `/levelrewards` to claim your level rewards.** \n## `/levelrewards`",
     requirementString: "Claim Level Rewards",
-    proceedString: "Use `/levelrewards` to claim your level rewards!",
+    proceedString:
+      "Use `/levelrewards` to claim your level rewards! \n## `/levelrewards`",
     checkRequirements: async (trainer) =>
       trainer.claimedLevelRewards.length > 0,
     rewards: {
@@ -341,10 +354,10 @@ const newTutorialConfigRaw = {
     name: "Purchasing Locations",
     emoji: "🌍",
     description:
-      "You may purchase and upgrade locations from the `/pokemart` for permanent boosts! **Use `/pokemart` to purchase the 🏠 Home location and upgrade it to level 3.** The Home locations improves the EXP gained from `/train`!\n\nYou can use `/locations` to view your locations.",
+      "You may purchase and upgrade locations from the `/pokemart` for permanent boosts! **Use `/pokemart` to purchase the 🏠 Home location and upgrade it to level 3.** The Home locations improves the EXP gained from `/train`!\n## `/pokemart`",
     requirementString: "Have a level 3 🏠 Home location",
     proceedString:
-      "Use `/pokemart` to purchase the Home location and upgrade it to level 3!",
+      "Use `/pokemart` to purchase the Home location and upgrade it to level 3! \n## `/pokemart`",
     checkRequirements: async (trainer) =>
       (trainer.locations?.[locations.HOME] ?? 0) >= 3,
     rewards: {
@@ -356,6 +369,7 @@ const newTutorialConfigRaw = {
     },
     image:
       "https://raw.githubusercontent.com/ewei068/pokestar/main/media/images/tutorial/home.png",
+    tip: "You can use /locations to view your locations",
   },
   beginnerLeveling: {
     name: "Beginner Pokemon Leveling",
@@ -363,7 +377,8 @@ const newTutorialConfigRaw = {
     description:
       "Now that you can `/train` your Pokemon faster, **train 6 Pokemon to level 15.**\n\nTip: On desktop, you can use the Up Arrow key to use your last command.",
     requirementString: "Train 6x Pokemon to level 15",
-    proceedString: "Use `/train` to train 6 Pokemon to level 15!",
+    proceedString:
+      "Use `/train` to train 6 Pokemon to level 15! \n## `/train <id>`",
     checkRequirements: async (trainer) => {
       const { data: pokemons } = await listPokemons(trainer, {
         pageSize: 6,
@@ -415,10 +430,10 @@ const newTutorialConfigRaw = {
     name: "Events",
     emoji: "🎉",
     description:
-      "Pokestar has events which provide **limited-time rewards, benefits, and custom Pokemon!** Use `/events` to view ongoing events.",
+      "Pokestar has events which provide **limited-time rewards, benefits, and custom Pokemon!** Use `/events` to view ongoing events. \n## `/events`",
     requirementString: "Complete the previous stage",
     proceedString:
-      "Use `/events` to view ongoing events, and complete the previous stage.",
+      "Use `/events` to view ongoing events, and complete the previous stage. \n## `/events`",
     checkRequirements: async (trainer) =>
       trainer.tutorialData.completedTutorialStages.winEasyDifficulty,
     rewards: {
@@ -447,10 +462,10 @@ const newTutorialConfigRaw = {
     name: "Settings",
     emoji: "⚙️",
     description:
-      "You can view and change your settings using `/settings`. **Use `/settings` to view and change your settings.**\n\nIf you're getting the hang of battles, you can speed them up by disabling target indicators. You can also set your device type to mobile for mobile optimizations.",
+      "You can view and change your settings using `/settings`. **Use `/settings` to view and change your settings.**\n\nIf you're getting the hang of battles, you can speed them up by disabling target indicators. You can also set your device type to mobile for mobile optimizations. \n## `/settings`",
     requirementString: "Complete the previous stage",
     proceedString:
-      "Use `/settings` to view and change your settings, and complete the previous stage.",
+      "Use `/settings` to view and change your settings, and complete the previous stage. \n## `/settings`",
     checkRequirements: async (trainer) =>
       trainer.tutorialData.completedTutorialStages.events,
     rewards: {
@@ -482,12 +497,13 @@ const newTutorialConfigRaw = {
         },
       },
     },
+    tip: "Battling will give you more EXP than training",
   },
   evolvePokemon: {
     name: "Evolving Pokemon",
     emoji: "🧬",
     description:
-      "Some Pokemon can evolve at the right level! Use `/pokedex` and click the growth tab to know when a Pokemon evolves, then **use `/evolve` to evolve a Pokemon.**",
+      "Some Pokemon can evolve at the right level! Use `/pokedex` and click the growth tab to know when a Pokemon evolves, then **use `/evolve` to evolve a Pokemon.** \n## `/pokedex <name>` `/evolve <name>`",
     requirementString: "Complete the previous stage",
     proceedString:
       "Use `/evolve` to evolve a Pokemon, and complete the previous stage.",
@@ -503,9 +519,10 @@ const newTutorialConfigRaw = {
     name: "Win Medium Difficulty",
     emoji: "🏆",
     description:
-      "Now that you have evolved your Pokemon, **use `/pve` to win a battle on Medium difficulty.**",
+      "Now that you have evolved your Pokemon, **use `/pve` to win a battle on Medium difficulty.** \n## `/pve`",
     requirementString: "Win any battle on Medium difficulty",
-    proceedString: "Use `/pve` to win a battle on Medium difficulty!",
+    proceedString:
+      "Use `/pve` to win a battle on Medium difficulty! \n## `/pve`",
     checkRequirements: async (trainer) => {
       for (const defeatedDifficulties of Object.values(trainer.defeatedNPCs)) {
         if (defeatedDifficulties.includes(difficulties.MEDIUM)) {
@@ -529,9 +546,9 @@ const newTutorialConfigRaw = {
     name: "Catching More Pokemon",
     emoji: emojis.ULTRABALL,
     description:
-      "You should keep using the `/gacha` to catch strong Pokemon. **Catch 50 Pokemon.**",
+      "You should keep using the `/gacha` to catch strong Pokemon. **Catch 50 Pokemon.** \n## `/gacha`",
     requirementString: "Catch 50x Pokemon",
-    proceedString: "Use `/gacha` to catch 50 Pokemon!",
+    proceedString: "Use `/gacha` to catch 50 Pokemon! \n## `/gacha`",
     checkRequirements: async (trainer) =>
       await checkNumPokemon(trainer)
         .then((res) => (res.numPokemon ?? 0) >= 50)
@@ -551,9 +568,10 @@ const newTutorialConfigRaw = {
     name: "Catching Darkrai",
     emoji: "<:491:1351027502801354822>",
     description:
-      "Darkrai is a Mythical Pokemon that can be caught after **completing 25 tutorial stages.** Use **`/mythic darkrai`** to catch Darkrai!",
+      "Darkrai is a Mythical Pokemon that can be caught after **completing 25 tutorial stages.** Use **`/mythic darkrai`** to catch Darkrai! \n## `/mythic darkrai`",
     requirementString: "Complete 25x tutorial stages and catch Darkrai",
-    proceedString: "Use `/mythic darkrai` to catch Darkrai!",
+    proceedString:
+      "Use `/mythic darkrai` to catch Darkrai! \n## `/mythic darkrai`",
     checkRequirements: async (trainer) => trainer.hasDarkrai,
     rewards: {
       money: 10000,
@@ -581,6 +599,7 @@ const newTutorialConfigRaw = {
     },
     image:
       "https://raw.githubusercontent.com/ewei068/pokestar/refs/heads/main/media/images/tutorial/auto-battle.png",
+    tip: "You can increase your Dream Card limit by leveling up your trainer",
   },
   advancedLeveling: {
     name: "Advanced Pokemon Leveling",
@@ -607,6 +626,7 @@ const newTutorialConfigRaw = {
         },
       },
     },
+    tip: "Battling will give you more EXP than training. Use auto battle to save time!",
   },
   winHardDifficulty: {
     name: "Win Hard Difficulty",