+++ changelog.md
@@ -2,6 +2,14 @@
 
 ## TODO
 
+- Improve UI
+  - Improve look of fields, use inline code when appropriate, use quotes to indent, use emojis!
+  - Also put more emojis in buttons
+  - Move, ability, and held item description formatting?
+- Put tutorial button in upsells
+  - Make the tutorial commands big asf
+  - Can add tips as footers
+
 **Stretch**
 
 - Stretch: Quests (Achievements)
@@ -12,6 +20,10 @@
 - Added new `/party manage` command! Much more easily add, move, and remove Pokemon from your party!
   - Adding pokemon from `/info` brings up the party manage UI
   - Added party manage to the `/tutorial`
+- Tutorial and general UI improvements
+  - Improved the UI of the `/tutorial` and other UI
+  - Re-ordered some tutorial stages
+  - Made the beginner gacha boost (10 pulls) apply to all banners
 
 ### 1.4.2 Gen 4 Part 2 (5/26)